@@ -3,37 +3,68 @@
   "version": "1.0.0",
   "license": "BSD-3-Clause",
   "scripts": {
-    "build": "yarn build:theme-prod && yarn build:docs",
+    "build": "run-s clean build:theme-prod build:docs build:pre-render",
     "build:docs": "jupyter-book build .",
-    "build:theme": "webpack build",
-    "build:theme-prod": "NODE_ENV=production yarn build:theme",
+    "build:theme": "vite build -m development",
+    "build:pre-render": "run-s build:pre-render:compile build:pre-render:types build:pre-render:rename-files build:pre-render:run",
+    "build:pre-render:compile": "PRE_RENDER=true vite build --ssr theme/src/renderAppToString.tsx --outDir ../renderAppToString/",
+    "build:pre-render:types": "tsc -p tsconfig.types.json",
+    "build:pre-render:rename-files": "mv renderAppToString/{renderAppToString,index}.js && mv renderAppToString/{renderAppToString,index}.d.ts",
+    "build:pre-render:run": "ts-node -P tsconfig.node.json -r tsconfig-paths/register pre-render",
+    "build:theme-prod": "vite build",
     "clean": "run-p clean:*",
     "clean:api-stable": "find api/stable/ ! -name 'index.rst' -type f -exec rm -f {} +",
     "clean:docs": "rm -rf _build/html",
+    "clean:pre-render": "rm -rf renderAppToString",
     "clean:theme": "rm -rf theme/napari/static/dist",
     "lint": "run-p lint:*",
-    "lint:scss": "stylelint theme/scss",
-    "lint:js": "eslint theme/js",
-    "dev": "yarn build:theme && mkdir -p _build/html && run-p dev:*",
-    "dev:server": "node dev-server",
-    "dev:theme": "webpack watch",
+    "lint:scss": "stylelint theme/src/scss",
+    "lint:ts": "eslint theme/src",
+    "dev": "mkdir -p _build/html && run-p dev:*",
+    "dev:server": "ts-node -P tsconfig.node.json dev-server",
+    "dev:theme": "vite",
+    "dev:tsm": "tsm -w -e default -p.@ theme/src 'theme/src/**/*.module.scss'",
     "postinstall": "husky install",
     "prettier": "prettier --check theme",
     "prettier:fix": "prettier --write theme"
   },
   "dependencies": {
-    "modern-normalize": "^1.0.0"
+    "@types/lodash-es": "^4.17.4",
+    "clsx": "^1.1.1",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "modern-normalize": "^1.0.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "slug": "^5.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^10.2.5",
+    "@tailwindcss/typography": "^0.4.1",
+    "@types/express": "^4.17.13",
+    "@types/morgan": "^1.9.3",
+    "@types/react": "^17.0.19",
+    "@types/react-dom": "^17.0.9",
+    "@types/slug": "^5.0.2",
+    "@typescript-eslint/eslint-plugin": "^4.21.0",
+    "@typescript-eslint/parser": "^4.29.3",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "autoprefixer": "^10.3.3",
+    "cheerio": "^1.0.0-rc.10",
     "chokidar": "^3.5.1",
     "css-loader": "^5.2.4",
     "cssnano": "^5.0.0-rc.2",
-    "eslint": "^7.22.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.3.1",
+    "eslint": "7.23.0",
+    "eslint-config-airbnb": "18.2.1",
+    "eslint-config-airbnb-typescript": "12.3.1",
+    "eslint-config-prettier": "8.1.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jest": "24.3.4",
+    "eslint-plugin-jest-playwright": "0.2.1",
+    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-prettier": "3.3.1",
+    "eslint-plugin-react": "7.23.2",
+    "eslint-plugin-react-hooks": "4.2.0",
+    "eslint-plugin-simple-import-sort": "7.0.0",
     "express": "^4.17.1",
     "fibers": "^5.0.0",
     "husky": "^5.2.0",
