@@ -1,47 +1,32 @@
-/**
- * Script for re-running `jupyter-book build` on script changes.
- *
- * TODO There's probably a smarter way to serve the re-built SCSS and JS assets
- * using some proxying magic. Instead of rebuilding the entire docs for SCSS
- * and JS changes, we could probably serve the re-compiled assets directly from
- * `theme/napari/static/dist` instead of rebuilding and using the updated
- * assets in `_build/html/_static/dist`.
- */
-
-const chokidar = require('chokidar');
-const { spawn } = require('child_process');
-const { resolve } = require('path');
-
-const express = require('express');
-const logger = require('morgan');
-const reload = require('reload');
+import { ChildProcess, spawn } from 'child_process';
+import chokidar from 'chokidar';
+import express, { Express } from 'express';
+import logger from 'morgan';
+import { resolve } from 'path';
+import reload, { Reloader } from 'reload';
 
 const PUBLIC_DIR = resolve(__dirname, '_build/html');
 const OUTPUT_DIR = resolve(__dirname, 'theme/napari/static/dist');
 const DEFAULT_PORT = 8080;
-const IGNORED_FILES = [
-  '.github',
-  '_build',
-  'node_modules',
-  'theme/js',
-  'theme/scss',
-];
+const IGNORED_FILES = ['.github', '_build', 'node_modules', 'theme/src'];
 
 class DevServer {
-  app = null;
-  buildProcess = null;
-  reloader = null;
+  app?: Express;
 
-  reloadBrowser() {
+  buildProcess?: ChildProcess;
+
+  reloader?: Reloader;
+
+  private reloadBrowser() {
     console.log('Reloading the browser!');
-    this.reloader.reload();
+    this.reloader?.reload();
   }
 
   /**
    * Runs `jupyter-book` to build the napari docs as a subprocess. Running
    * processes are cancelled to ensure the build is up-to-date.
    */
-  buildDocs() {
+  private buildDocs() {
     // Kill existing build if there is one
     if (this.buildProcess) {
       console.log('Cancelling existing build');
@@ -53,7 +38,7 @@ class DevServer {
       stdio: 'inherit',
     });
     this.buildProcess.once('exit', (code) => {
-      this.buildProcess = null;
+      this.buildProcess = undefined;
 
       // Reload on successful run
       if (this.reloader && code === 0) {
@@ -65,7 +50,7 @@ class DevServer {
   /**
    * Re-builds the napari docs on file changes.
    */
-  watchDocs() {
+  private watchDocs() {
     const watcher = chokidar.watch('.', { ignored: IGNORED_FILES });
 
     watcher.on('change', (path) => {
@@ -87,12 +72,12 @@ class DevServer {
   /**
    * Starts the dev server using express and reload.
    */
-  async startDevServer() {
+  private async startDevServer() {
     this.app = express();
     this.app.use(logger('dev'));
 
-    // Proxy static path used by sphinx to bypass having to rebuild the docs
-    // when the JS / SCSS changes.
+    // Proxy static path used by Jupyter Book to bypass having to rebuild the
+    // docs when the JS / SCSS changes.
     this.app.use('/_static/dist', express.static(OUTPUT_DIR));
 
     // Serve built docs if other proxy handler is a miss.
@@ -113,4 +98,6 @@ class DevServer {
 }
 
 const server = new DevServer();
-server.start();
+server
+  .start()
+  .catch((err) => console.error('Unable to start dev server!', err));
