@@ -1,6 +1,7 @@
 module.exports = {
   plugins: {
-    'postcss-import': {},
+    tailwindcss: {},
     autoprefixer: {},
+    'postcss-import': {},
   },
 };
